+++ jlrs/tests/async_util/async_tasks.rs
@@ -1,15 +1,18 @@
-use jlrs::{async_util::task::Register, memory::gc::Gc, prelude::*};
+use jlrs::{
+    async_util::task::{AsyncTask, Register},
+    memory::gc::Gc,
+    prelude::*,
+};
 
 pub struct MyTask {
     pub dims: isize,
     pub iters: isize,
 }
 
-#[async_trait(?Send)]
 impl AsyncTask for MyTask {
     type Output = JlrsResult<f64>;
 
-    async fn run<'base>(&mut self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
+    async fn run<'base>(self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
         let dims = Value::new(&mut frame, self.dims);
         let iters = Value::new(&mut frame, self.iters);
 
@@ -37,11 +40,10 @@ pub struct OtherRetTypeTask {
     pub iters: isize,
 }
 
-#[async_trait(?Send)]
 impl AsyncTask for OtherRetTypeTask {
     type Output = f32;
 
-    async fn run<'base>(&mut self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
+    async fn run<'base>(self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
         let dims = Value::new(&mut frame, self.dims);
         frame.gc_collect_n(jlrs::memory::gc::GcCollection::Full, 3);
 
@@ -72,12 +74,10 @@ pub struct KwTask {
     pub iters: isize,
 }
 
-#[async_trait(?Send)]
 impl AsyncTask for KwTask {
     type Output = JlrsResult<f32>;
 
-    async fn run<'base>(&mut self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
-        // println!("KwTask {:p}", frame.stack_addr());
+    async fn run<'base>(self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
         let dims = Value::new(&mut frame, self.dims);
         frame.gc_collect_n(jlrs::memory::gc::GcCollection::Full, 3);
         let iters = Value::new(&mut frame, self.iters);
@@ -99,7 +99,6 @@ impl AsyncTask for KwTask {
                 .unwrap()
                 .unbox::<f64>()? as f32
         };
-        // println!("KwTask done");
 
         Ok(v)
     }
@@ -107,12 +106,10 @@ impl AsyncTask for KwTask {
 
 pub struct ThrowingTask;
 
-#[async_trait(?Send)]
 impl AsyncTask for ThrowingTask {
     type Output = JlrsResult<f32>;
 
-    async fn run<'base>(&mut self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
-        // println!("ThrowingTask {:p}", frame.stack_addr());
+    async fn run<'base>(self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
         frame.gc_collect_n(jlrs::memory::gc::GcCollection::Full, 3);
 
         let v = unsafe {
@@ -126,7 +123,6 @@ impl AsyncTask for ThrowingTask {
                 .into_jlrs_result()?
                 .unbox::<f64>()? as f32
         };
-        // println!("ThrowingTask done");
 
         Ok(v)
     }
@@ -137,12 +133,10 @@ pub struct NestingTaskAsyncFrame {
     pub iters: isize,
 }
 
-#[async_trait(?Send)]
 impl AsyncTask for NestingTaskAsyncFrame {
     type Output = JlrsResult<f64>;
 
-    async fn run<'base>(&mut self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
-        // println!("NestingTaskAsyncFrame {:p}", frame.stack_addr());
+    async fn run<'base>(self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
         let dims = Value::new(&mut frame, self.dims);
         frame.gc_collect_n(jlrs::memory::gc::GcCollection::Full, 3);
         let iters = Value::new(&mut frame, self.iters);
@@ -168,7 +162,6 @@ impl AsyncTask for NestingTaskAsyncFrame {
             .await?;
 
         frame.gc_collect_n(jlrs::memory::gc::GcCollection::Full, 3);
-        // println!("NestingTaskAsyncFrame done");
 
         Ok(v)
     }
@@ -179,19 +172,16 @@ pub struct NestingTaskAsyncValueFrame {
     pub iters: isize,
 }
 
-#[async_trait(?Send)]
 impl AsyncTask for NestingTaskAsyncValueFrame {
     type Output = JlrsResult<f64>;
 
-    async fn run<'base>(&mut self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
-        // println!("NestingTaskAsyncValueFrame {:p}", frame.stack_addr());
+    async fn run<'base>(self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
         let output = frame.output();
 
         frame.gc_collect_n(jlrs::memory::gc::GcCollection::Full, 3);
 
         let v = (&mut frame)
             .async_scope(|mut frame| async move {
-                // println!("NestingTaskAsyncFrame");
                 let iters = Value::new(&mut frame, self.iters);
 
                 frame.gc_collect_n(jlrs::memory::gc::GcCollection::Full, 3);
@@ -217,7 +207,6 @@ impl AsyncTask for NestingTaskAsyncValueFrame {
             })
             .await?
             .unbox::<f64>()?;
-        // println!("NestingTaskAsyncValueFrame done");
 
         Ok(v)
     }
@@ -228,12 +217,10 @@ pub struct NestingTaskAsyncCallFrame {
     pub iters: isize,
 }
 
-#[async_trait(?Send)]
 impl AsyncTask for NestingTaskAsyncCallFrame {
     type Output = JlrsResult<f64>;
 
-    async fn run<'base>(&mut self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
-        // println!("NestingTaskAsyncCallFrame {:p}", frame.stack_addr());
+    async fn run<'base>(self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
         let output = frame.output();
         frame.gc_collect_n(jlrs::memory::gc::GcCollection::Full, 3);
 
@@ -269,7 +256,6 @@ impl AsyncTask for NestingTaskAsyncCallFrame {
             .await?
             .unwrap()
             .unbox::<f64>()?;
-        // println!("NestingTaskAsyncCallFrame done");
 
         Ok(v)
     }
@@ -280,12 +266,10 @@ pub struct NestingTaskAsyncGcFrame {
     pub iters: isize,
 }
 
-#[async_trait(?Send)]
 impl AsyncTask for NestingTaskAsyncGcFrame {
     type Output = JlrsResult<f64>;
 
-    async fn run<'base>(&mut self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
-        // println!("NestingTaskAsyncGcFrame {:p}", frame.stack_addr());
+    async fn run<'base>(self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
         let dims = Value::new(&mut frame, self.dims);
         frame.gc_collect_n(jlrs::memory::gc::GcCollection::Full, 3);
         let iters = Value::new(&mut frame, self.iters);
@@ -310,7 +294,6 @@ impl AsyncTask for NestingTaskAsyncGcFrame {
                 }
             })
             .await?;
-        // println!("NestingTaskAsyncGcFrame done");
 
         Ok(v)
     }
@@ -321,12 +304,10 @@ pub struct NestingTaskAsyncDynamicValueFrame {
     pub iters: isize,
 }
 
-#[async_trait(?Send)]
 impl AsyncTask for NestingTaskAsyncDynamicValueFrame {
     type Output = JlrsResult<f64>;
 
-    async fn run<'base>(&mut self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
-        // println!("NestingTaskAsyncDynamicValueFrame {:p}", frame.stack_addr());
+    async fn run<'base>(self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
         let output = frame.output();
         let v = frame
             .async_scope(|mut frame| async move {
@@ -350,12 +331,10 @@ impl AsyncTask for NestingTaskAsyncDynamicValueFrame {
                         .unwrap()
                 };
 
-                // println!("Root {out:?}");
                 Ok(out.root(output))
             })
             .await?
             .unbox::<f64>()?;
-        // println!("NestingTaskAsyncDynamicValueFrame done");
 
         Ok(v)
     }
@@ -366,12 +345,10 @@ pub struct NestingTaskAsyncDynamicCallFrame {
     pub iters: isize,
 }
 
-#[async_trait(?Send)]
 impl AsyncTask for NestingTaskAsyncDynamicCallFrame {
     type Output = JlrsResult<f64>;
 
-    async fn run<'base>(&mut self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
-        // println!("NestingTaskAsyncDynamicCallFrame {:p}", frame.stack_addr());
+    async fn run<'base>(self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
         let output = frame.output();
         let v = frame
             .async_scope(|mut frame| async move {
@@ -416,7 +393,6 @@ impl AsyncTask for NestingTaskAsyncDynamicCallFrame {
             .await?
             .unwrap()
             .unbox::<f64>()?;
-        // println!("NestingTaskAsyncDynamicCallFrame done");
 
         Ok(v)
     }
@@ -426,7 +402,6 @@ pub struct AccumulatorTask {
     pub init_value: f64,
 }
 
-#[async_trait(?Send)]
 impl Register for AccumulatorTask {
     async fn register<'frame>(mut frame: AsyncGcFrame<'frame>) -> JlrsResult<()> {
         unsafe {
@@ -437,7 +412,6 @@ impl Register for AccumulatorTask {
     }
 }
 
-#[async_trait(?Send)]
 impl PersistentTask for AccumulatorTask {
     type State<'state> = Value<'state, 'static>;
     type Input = f64;
@@ -449,7 +423,6 @@ impl PersistentTask for AccumulatorTask {
         &mut self,
         mut frame: AsyncGcFrame<'frame>,
     ) -> JlrsResult<Value<'frame, 'static>> {
-        // println!("AccumulatorTask init {:p}", frame.stack_addr());
         unsafe {
             let output = frame.output();
             let init_value = self.init_value;
@@ -471,7 +444,6 @@ impl PersistentTask for AccumulatorTask {
                 .await?
                 .into_jlrs_result();
 
-            // println!("AccumulatorTask init");
             res
         }
     }
@@ -482,7 +454,6 @@ impl PersistentTask for AccumulatorTask {
         state: &mut Self::State<'state>,
         input: Self::Input,
     ) -> Self::Output {
-        // println!("AccumulatorTask run {:p}", frame.stack_addr());
         let value = state.field_accessor().field("v")?.access::<f64>()? + input;
         let new_value = Value::new(&mut frame, value);
 
@@ -495,7 +466,6 @@ impl PersistentTask for AccumulatorTask {
         frame.gc_collect(jlrs::memory::gc::GcCollection::Full);
         frame.gc_collect(jlrs::memory::gc::GcCollection::Full);
         frame.gc_collect(jlrs::memory::gc::GcCollection::Full);
-        // println!("AccumulatorTask run done");
 
         Ok(value)
     }
@@ -506,12 +476,10 @@ pub struct LocalTask {
     pub iters: isize,
 }
 
-#[async_trait(?Send)]
 impl AsyncTask for LocalTask {
     type Output = JlrsResult<f32>;
 
-    async fn run<'base>(&mut self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
-        // println!("LocalTask {:p}", frame.stack_addr());
+    async fn run<'base>(self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
         let dims = Value::new(&mut frame, self.dims);
         let iters = Value::new(&mut frame, self.iters);
 
@@ -530,7 +498,6 @@ impl AsyncTask for LocalTask {
         frame.gc_collect(jlrs::memory::gc::GcCollection::Full);
         frame.gc_collect(jlrs::memory::gc::GcCollection::Full);
         frame.gc_collect(jlrs::memory::gc::GcCollection::Full);
-        // println!("LocalTask done");
 
         Ok(v)
     }
@@ -541,12 +508,10 @@ pub struct LocalSchedulingTask {
     pub iters: isize,
 }
 
-#[async_trait(?Send)]
 impl AsyncTask for LocalSchedulingTask {
     type Output = JlrsResult<f32>;
 
-    async fn run<'base>(&mut self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
-        // println!("LocalSchedulingTask {:p}", frame.stack_addr());
+    async fn run<'base>(self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
         let dims = Value::new(&mut frame, self.dims);
         let iters = Value::new(&mut frame, self.iters);
 
@@ -570,8 +535,6 @@ impl AsyncTask for LocalSchedulingTask {
                 .unbox::<f64>()? as f32
         };
 
-        // println!("LocalSchedulingTask done");
-
         Ok(v)
     }
 }
@@ -581,12 +544,10 @@ pub struct MainTask {
     pub iters: isize,
 }
 
-#[async_trait(?Send)]
 impl AsyncTask for MainTask {
     type Output = JlrsResult<f32>;
 
-    async fn run<'base>(&mut self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
-        // println!("MainTask {:p}", frame.stack_addr());
+    async fn run<'base>(self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
         let dims = Value::new(&mut frame, self.dims);
         let iters = Value::new(&mut frame, self.iters);
 
@@ -605,7 +566,6 @@ impl AsyncTask for MainTask {
         frame.gc_collect(jlrs::memory::gc::GcCollection::Full);
         frame.gc_collect(jlrs::memory::gc::GcCollection::Full);
         frame.gc_collect(jlrs::memory::gc::GcCollection::Full);
-        // println!("MainTask done");
 
         Ok(v)
     }
@@ -616,12 +576,10 @@ pub struct MainSchedulingTask {
     pub iters: isize,
 }
 
-#[async_trait(?Send)]
 impl AsyncTask for MainSchedulingTask {
     type Output = JlrsResult<f32>;
 
-    async fn run<'base>(&mut self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
-        // println!("MainSchedulingTask {:p}", frame.stack_addr());
+    async fn run<'base>(self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
         let dims = Value::new(&mut frame, self.dims);
         let iters = Value::new(&mut frame, self.iters);
 
@@ -644,7 +602,6 @@ impl AsyncTask for MainSchedulingTask {
                 .into_jlrs_result()?
                 .unbox::<f64>()? as f32
         };
-        // println!("MainSchedulingTask done");
 
         Ok(v)
     }
@@ -655,12 +612,10 @@ pub struct SchedulingTask {
     pub iters: isize,
 }
 
-#[async_trait(?Send)]
 impl AsyncTask for SchedulingTask {
     type Output = JlrsResult<f32>;
 
-    async fn run<'base>(&mut self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
-        // println!("SchedulingTask {:p}", frame.stack_addr());
+    async fn run<'base>(self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
         let dims = Value::new(&mut frame, self.dims);
         let iters = Value::new(&mut frame, self.iters);
 
@@ -684,7 +639,6 @@ impl AsyncTask for SchedulingTask {
                 .unbox::<f64>()? as f32
         };
 
-        // println!("SchedulingTask done");
         Ok(v)
     }
 }
@@ -694,12 +648,10 @@ pub struct LocalKwSchedulingTask {
     pub iters: isize,
 }
 
-#[async_trait(?Send)]
 impl AsyncTask for LocalKwSchedulingTask {
     type Output = JlrsResult<f32>;
 
-    async fn run<'base>(&mut self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
-        // println!("LocalKwSchedulingTask {:p}", frame.stack_addr());
+    async fn run<'base>(self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
         let dims = Value::new(&mut frame, self.dims);
         let iters = Value::new(&mut frame, self.iters);
 
@@ -726,7 +678,6 @@ impl AsyncTask for LocalKwSchedulingTask {
                 .into_jlrs_result()?
                 .unbox::<f64>()? as f32
         };
-        // println!("LocalKwSchedulingTask done");
 
         Ok(v)
     }
@@ -737,12 +688,10 @@ pub struct KwSchedulingTask {
     pub iters: isize,
 }
 
-#[async_trait(?Send)]
 impl AsyncTask for KwSchedulingTask {
     type Output = JlrsResult<f32>;
 
-    async fn run<'base>(&mut self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
-        // println!("KwSchedulingTask {:p}", frame.stack_addr());
+    async fn run<'base>(self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
         let dims = Value::new(&mut frame, self.dims);
         let iters = Value::new(&mut frame, self.iters);
 
@@ -769,7 +718,6 @@ impl AsyncTask for KwSchedulingTask {
                 .into_jlrs_result()?
                 .unbox::<f64>()? as f32
         };
-        // println!("KwSchedulingTask done");
 
         Ok(v)
     }
@@ -780,12 +728,10 @@ pub struct MainKwSchedulingTask {
     pub iters: isize,
 }
 
-#[async_trait(?Send)]
 impl AsyncTask for MainKwSchedulingTask {
     type Output = JlrsResult<f32>;
 
-    async fn run<'base>(&mut self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
-        // println!("MainKwSchedulingTask {:p}", frame.stack_addr());
+    async fn run<'base>(self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
         let dims = Value::new(&mut frame, self.dims);
         let iters = Value::new(&mut frame, self.iters);
 
@@ -812,7 +758,6 @@ impl AsyncTask for MainKwSchedulingTask {
                 .into_jlrs_result()?
                 .unbox::<f64>()? as f32
         };
-        // println!("MainKwSchedulingTask done");
 
         Ok(v)
     }
@@ -823,12 +768,10 @@ pub struct LocalKwTask {
     pub iters: isize,
 }
 
-#[async_trait(?Send)]
 impl AsyncTask for LocalKwTask {
     type Output = JlrsResult<f32>;
 
-    async fn run<'base>(&mut self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
-        // println!("LocalKwTask {:p}", frame.stack_addr());
+    async fn run<'base>(self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
         let dims = Value::new(&mut frame, self.dims);
         let iters = Value::new(&mut frame, self.iters);
         let kw = Value::new(&mut frame, 5.0f64);
@@ -850,7 +793,6 @@ impl AsyncTask for LocalKwTask {
                 .unwrap()
                 .unbox::<f64>()? as f32
         };
-        // println!("LocalKwTask done");
 
         Ok(v)
     }
@@ -861,12 +803,10 @@ pub struct MainKwTask {
     pub iters: isize,
 }
 
-#[async_trait(?Send)]
 impl AsyncTask for MainKwTask {
     type Output = JlrsResult<f32>;
 
-    async fn run<'base>(&mut self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
-        // println!("MainKwTask {:p}", frame.stack_addr());
+    async fn run<'base>(self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
         let dims = Value::new(&mut frame, self.dims);
         let iters = Value::new(&mut frame, self.iters);
         let kw = Value::new(&mut frame, 5.0f64);
@@ -888,7 +828,6 @@ impl AsyncTask for MainKwTask {
                 .unwrap()
                 .unbox::<f64>()? as f32
         };
-        // println!("MainKwTask done");
 
         Ok(v)
     }
@@ -896,12 +835,10 @@ impl AsyncTask for MainKwTask {
 
 pub struct BorrowArrayData;
 
-#[async_trait(?Send)]
 impl AsyncTask for BorrowArrayData {
     type Output = JlrsResult<f64>;
 
-    async fn run<'base>(&mut self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
-        // println!("BorrowArrayData {:p}", frame.stack_addr());
+    async fn run<'base>(self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
         let mut data = vec![2.0f64];
         let borrowed = &mut data;
         let output = frame.output();
@@ -920,7 +857,6 @@ impl AsyncTask for BorrowArrayData {
         // Uncommenting next line must be compile error
         // let _ = data[0];
         let v = data2[0];
-        // println!("BorrowArrayData done");
         Ok(v)
     }
 }
@@ -931,12 +867,10 @@ impl AsyncTask for BorrowArrayData {
 
 // TODO: Rust 1.85
 // #[cfg(feature = "async-closure")]
-// #[async_trait(?Send)]
 // impl AsyncTask for BorrowArrayDataClosure {
 //     type Output = JlrsResult<f64>;
 
-//     async fn run<'base>(&mut self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
-//         // println!("BorrowArrayData {:p}", frame.stack_addr());
+//     async fn run<'base>(self, mut frame: AsyncGcFrame<'base>) -> Self::Output {
 //         let mut data = vec![2.0f64];
 //         let borrowed = &mut data;
 //         let output = frame.output();
@@ -953,7 +887,6 @@ impl AsyncTask for BorrowArrayData {
 //         // Uncommenting next line must be compile error
 //         // let _ = data[0];
 //         let v = data2[0];
-//         // println!("BorrowArrayData done");
 //         Ok(v)
 //     }
 // }